python_sms_engine/requirements.txt
fastapi
uvicorn
pyserial
pydantic
